SCRUM-31 Generate and show config on button press
